Added a framework for storing options which are persistent for the life of the view. Currently there is no frontened way to change them, but testing them by changing the code indicates that the optiosn properly change the running of the command.
The next step will be to implement a frontend to be able to change the options.

Rather than going for a customized run option, I am going to go in the direction of a command which will change the options that the normal run uses. This later can be extended to pull the default options from existing configuration files.